requirements.txt
customtkinter
opencv-python
pillow
reportlab
tensorflow
scikit-learn
pandas
matplotlib